refactor(app) : util함수들의 export를 묶는다
