+++ packages/app/src/components/MainNavigator.tsx
@@ -8,7 +8,7 @@ import TabBarNavigator from './TabBarNavigator';
 import WebViewContainer from './WebViewContainer';
 
 import { useDidMount } from '@/hooks/useDidMount';
-import { setFcmAlert } from '@/utils/setFcmAlert';
+import { setFcmAlert } from '@/utils';
 
 const Stack = createStackNavigator();
